Module proj_billing::billing::consumption::integer_consumption

Integer Consumption

Billing Protocol allows protocols to define their own pricing strategy. This module implements an example pricing strategy: hourly time of use billing.

Structs

IntegerConsumption
Consumption information for hourly time of use billing

Type Definitions

Prices
Co-efficient for the number of consumption units for each hour of each day of the week