Module proj_billing::billing::sign_on_meter

Sign on meter

The simplest smart meter billing protocol. This implements linear time of use billing under the assumption that everything is run on tamperproof hardware in the smart meter. This is very unrealistic. The privacy issues relating to time of use billing are overcome by calculating the bill total on the meter and then only sending this total to the utility company every billing period (e.g. one month).

Structs

Consumption
Consumption information for hourly time of use billing

SignOnMeter
State of the billing protocol

Type Definitions

Prices
Co-efficient for the number of consumption units for each hour of each day of the week